requirements.txt
crewai[tools]==1.9.3
python-dotenv==1.1.1
streamlit==1.54.0
langchain-community==0.4.1
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
sentence-transformers==5.2.3
litellm==1.75.3
